system/rust: Refactor server connections/clients
Prior to this CL, AttDatabaseImpl actually represented a client
connection to a particular server. This CL is a refactor that changes it
so that we name it AttClient, which represents a particular connection
to a server and we make it point at the same object that is contained
within the collection of clients that is managed GattModule.

As part of this refactor, I have removed the AttDatabase trait which was
unnecessary. It only existed for testing, but we already have a trait
that provides the abstraction we need: RawGattDatastore.

This change is motivated by the need to implement the prepare-write and
execute-write ATT requests which are required to support EATT. To
implement those requests, we need access to some per-client state.

Bug: 397308818
Flag: EXEMPT refactor only
Test: atest --host libbluetooth_core_rs_test
Change-Id: I843ac923abb10109343a9dfc679355139016181a