Revert "GMCS: Fix for long read"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2c174ffb-5437-420e-a8e7-48f77cf3ea9b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib2c5bab3e13023b10d63cd50c39acd865b898214
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com