Revert "Flag for correctly tracking temporary pairing"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d6b32ae6-e5e8-4354-9586-113f05b46a99). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5a88d0e6c187e7810cf683baf24b3ded14618577
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com