le_audio: Update inband ringtone for fallback device
With this CL Inband ringtones are now handled for:
* Devices in the currently active group.
* Devices in the broadcast-to-unicast fallback group during an active broadcast.
* Devices in the broadcast-to-unicast fallback group when a broadcast is ready for activation.

The current implementation does not dynamically update inband ringtone support
during a broadcast-to-unicast fallback group change. Specifically, when the
fallback group transitions, the previous group's inband ringtone support is not
immediately revoked, and the new group's support is not immediately granted.

To address the expanded support, the following scenarios now trigger a
re-evaluation of inband ringtone support:
* Group activation.
* Group deactivation.
* Group configuration changes.
* Handover from unicast to broadcast mode.
* Changes to the broadcast-to-unicast fallback group while in broadcast mode.

Bug: 397946268
Test: atest LeAudioBroadcastServiceTest
Flag: EXEMPT, regression covered by unit tests
Change-Id: I1a3cd79e6224ccdc05b4122081c97f4c189b275a