Revert "Add flag scan_controller_thread"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1eb07085-ea16-46f5-b9d5-456e933921d8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9537d749ae26cfafd4327f42f9c3a4ae6a2c4bd1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com