Revert "Add custom tracing to bluetooth using Tracing SDK"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/545fd3c7-ea99-4201-b64a-3dc0f50bb5a2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I652982ee55bb7e82201cd165eeb6de7ac2b2f00c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com