Revert "system/gd: Remove deprecated testing documentation"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/75abb48a-875d-47d9-9e79-1e8ab291ae80). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idf3774fd75466390151f8d6038fefa007339e88e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com