Revert "leaudio: Inactivate device on autonomous stream release"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fd7722e7-135f-48ec-8285-696ac9cf10c1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3dfcfce5b7b0fbbd4786a66714fa8e32142386c6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com