le_audio: adding config for single bud connection
During connection with buds pair, if a single bud connects first,
and any use-case starts, SingleDev config was being selected.
After sometime when the second bud connected, DualDev config
was being selected. But the 'params' for DualDev config did not
match with that of SingleDev config, leading to BT crash.
In the fix, added corresponding SingleDev config for the required
DualDev config, so that when params are matched at ENABLE,
crash does not occur

Bug: 398977082
Test: mmm pacaktes/modules/Bluetooth
Flag: Exempt, day-to-day fix.

Change-Id: I5c245c1c595e08269f7c2beb3c2c94ff533b382b