Revert "floss: Fix the battery level update in the BAS"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0eb2f81d-043c-43db-9ba0-4378d50f8385). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I868641b60f60f66e6f676a009f1587760cf53e0d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com