Revert "Add flag scan_controller_thread"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/180b72c0-d2e0-45a9-9781-db5d8743dc0a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie5be212e00fd98671c56bdc41c65498cef47a3cd
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com