run bluetooth_test_gd_unit for system change only
The bluetooth_test_gd_unit is flaky. It is trigger unconditionally to
any change in the Bluetooth.
When modifying java code, this has no impact on bluetooth_test_gd_unit.
Same when adding flag or any other folder.

Since this is a gd test it is in theory possible to have a more narrow
window that trigger the test, but I don't want to assume too much

Bug: 311772251
Test: test mapping
Flag: Exempt TEST_ONLY
Change-Id: I295a02051633a8a78b8e7eb62d5fc53f8bbf7972