Revert "BluetoothMetrics: Timeout failure for LE-ACL Connection"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a921842a-3874-441e-af1b-4e2c58c28824). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic20779a0d9a6aa5d2b017fbd93a92928f3a4a2d8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com