Revert "InCallService: clccResponse stays within Bt"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/929dd8c0-62ba-4134-a9d4-52ab3c1aced7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic04cd1d7173fb54a219b0f057166589a349b87c9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com