Revert "Remove leaudio_broadcast_extract_periodic_scanner_from_s..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/929dd8c0-62ba-4134-a9d4-52ab3c1aced7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9f658cad58b175657d8714a61a18ad46ab9fdf18
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com