Revert "Errorprone fix & enforce MethodCanBeStatic"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/55f18f24-9f1a-405a-8cc8-0444de930a35). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 399256764
Change-Id: I6f7f021a6184c1faed40665cd4f9b56743d85d19
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com