Revert "Errorprone fix & enforce MethodCanBeStatic"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/739d98dc-7140-4316-b866-05ccb70799c7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 399219878
Change-Id: I950ae4aba7035aed28c87181f454fdda0963653b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com