leaudio: Change Release behavior on IDLE ASEs
With this change, Android will not send Release Ctp to the ASEs which
are in the IDLE to avoild IOP issue.

In case race will happen: remote did not notify Android about Codec
Configured state when Android was trying to send Release, then Codec
Configured notification will be treat as autonomous change of the remote
device and will not cause any futher actions.

Bug: 385283379
Test: atest bluetooth_le_audio_test
Flag: Exempt, verified with unit tests. Unit tests touching this
scenario got fixed.

Change-Id: I7e6c8aafa51c4874df99d331b2a0c85cc8016f86