leaudio: Change Release behavior on IDLE ASEs and handling CTP errors
With this change, Android will not send Release Ctp to the ASEs which
are in the IDLE to avoid IOP issue and will be smarter on Control point
error handling. Those two things we want to have in a single CL as we
cannot deliver one part of this only.

In case race on ASEs state happens i.e.: remote did not notify Android
about Codec Configured state when Android was trying to send Release, then the
future Codec Configured notification will be treat as autonomous
change of the remote device and will not cause any futher actions.

When error will happen on the Ctp message on a single device, this
device will get Released from the stream. If all set members will get an
error on the Ctp message, then all the devices will be release and user
will be notified about stream Realeasing.

Bug: 385283379
Test: atest bluetooth_le_audio_test
Flag: Exempt, verified with unit tests. Unit tests touching this
scenario got fixed.

Change-Id: I7e6c8aafa51c4874df99d331b2a0c85cc8016f86