Revert "floss: Fix the battery level update in the BAS"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5c713fe2-884f-4226-b35e-c0c3926588c9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I50b10b5db4bb6be5bbbbfcdc0b0a554163e9105d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com