Revert "Add flag scan_controller_thread"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ba024b84-e6cf-40e9-807f-8833c2d314fc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie293f62ad687ca342749675c3673000caaf9b9f0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com