Revert "VC: Extend RemovePendingVolumeControlOperations to accep..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7aeb88ad-d9dc-4cea-9841-ab497479af99). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6cf1039ddea9d6c5b1e6ebd97219b67f52f9696c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com