Revert "VC: Allow to mute/unmute if different from current or st..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7aeb88ad-d9dc-4cea-9841-ab497479af99). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iaa1f1a5d737938c00f0e2a66dec70f64462e5378
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com