Revert "VC: Allow to set volume if different from current or sta..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7aeb88ad-d9dc-4cea-9841-ab497479af99). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I35caf101292cac1fdbac20910bae098848e55497
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com