Revert "Add flag scan_controller_thread"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/030b762f-1877-427d-af8d-c53c7f79f75c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9e1c3db963e05c5977d77eb4dfcedd3778afbf9f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com