Revert "floss: Fix the battery level update in the BAS"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dba05a09-c52c-4a6b-8761-d0d75c1f9574). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibf8cd24143096109addaffe5b9f27a407b851339
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com