Revert "RAS: Set eatt_support false for BTA_GATTC_AppRegister"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/12b66c88-5b8d-41ab-b2d0-836c32e22094). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I15c84de1979db6c1aaa9c89b97a1414e271cad46
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com