Revert "floss: Fix the battery level update in the BAS"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/37ae41cd-53e5-4740-a006-0998bc75c8e5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic47afef420129db101b86760702e3a6bb78087c4
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com