Revert "BluetoothService: Respect config_hsumBootStrategy"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e368394a-5ff2-41f1-9778-75d65fcf219b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If48392b340887f436cd3dc608fb2ba2bd68623aa
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com