Revert "Fix typos in variable names, comments, and documentation..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4c4f858a-b89a-47a1-b191-7a97a664cda9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idd064fa93f0c274b7cad36358c2a064bad023bd5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com