Revert "Flag for allowing early incoming HID connection"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4c4f858a-b89a-47a1-b191-7a97a664cda9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0432229a7b8d1f7dfba784b449435f33aeb1bbba
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com