Revert "le_audio: Wait for all CISes and Data Paths to be complete"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bcd504fd-70e9-4ce1-972b-b2adfe408667). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If62ff18d66bcb2b1f4a4a12a8b91cc2243c45c63
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com