Revert "le_audio: state_machine: Ignore unexpected CIS establish..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bcd504fd-70e9-4ce1-972b-b2adfe408667). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I09d09fd957a19aec0122508c7a9066739370dc6f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com