Revert "a2dp/aidl: Fix missing codec_type in a2dp provider info"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f37eb74b-657d-42fd-aec0-ea255c5fd0d9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ice7e5e51e46e6b286548afbb41192b210dc596fd
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com