Revert "stack/a2dp_ext: Write the provider capabilities to codec..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f37eb74b-657d-42fd-aec0-ea255c5fd0d9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I75b9f18bd53aaa3c5da3cecae812aa4a4d529bfb
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com