Revert "stack/a2dp: Remove the field A2dpCodecConfig#codec_capab..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f37eb74b-657d-42fd-aec0-ea255c5fd0d9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibfdbf645d510dac87495113e4ceb7d611879e841
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com