Revert "[owners] Remove eruffieux@google.com from apex/OWNERS"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f37eb74b-657d-42fd-aec0-ea255c5fd0d9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If265b6219053063755e6074fd69c51b91f2aad7e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com