Revert "floss: Fix the battery level update in the BAS"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/18dad588-6e2a-456a-81f7-d02c258d1dfb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2324d9e0b7a1c14e3500cff4c5f3dafb42b92456
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com