Revert "le_audio: state_machine: Ignore unexpected CIS establish..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d60b5d3b-6ae5-426a-a049-bda944877700). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia800330ced928db7e466a3e6752683579f84d9fe
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com