Revert "le_audio: Wait for all CISes and Data Paths to be complete"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d60b5d3b-6ae5-426a-a049-bda944877700). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic236351b4098972980290295c28007e3fd8929b7
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com