Revert "Move CounterMetrics out of the module registry"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/32b6ccb1-d4dd-47d6-9692-85f2733146d4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ief06ed382adae004d4dfd4a87c62d5e9fe6e1568
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com