Revert "bluetooth_test_gd_unit go to the corner"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/664ee04d-b901-4060-a76d-9c351c9b56ee). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibf3d7fda76ec12f419929df257d8eeeed5b45aa9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com