Revert "le_audio: Don't QoS configure if streaming stopped"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/664ee04d-b901-4060-a76d-9c351c9b56ee). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iaa530f26e16a85c80f9bc0c4c65839214fffb2cf
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com