Revert "Remove dependency from system/stack/test on system/gd"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/db525c66-16c8-48c4-8ee9-4d79a40e4a9a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id0216f1ffa41cf881d48345c0bca664a697e65a4
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com