Revert "Flag for allowing early incoming HID connection"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5805d36b-494a-4df8-890e-cb2fced2d865). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7e20dd017b8f921c3e6532830224e1640ec43768
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com