Revert "re-activate PbapStateMachineTest with proper mock"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c394b089-5e3d-4802-836d-d985aa592f04). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I295e54e5e918effe5f4b87372a326a96f9654a80
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com