Revert "avdt: add timer for monitoring delay report as initiator"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cb342be8-abff-42bb-b878-bfb100882d6a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I073e2269910afe3c0371561ba793228c34e40875
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com