Revert "Add flag scan_controller_thread"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/88ca6aa3-a98d-4650-8562-cf5e87da69e2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I16f4b12d553bd03b5e30ff795ff76eddaf3681ef
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com