Revert "Cleanup bp files in system/stack/test/fuzzers"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9d14ad0b-fab7-402b-aa16-c565915639b3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iafc644747538477723b53ad360377b9b1ff88292
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com