Revert "Remove dependency from system/stack/test on system/gd"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0b453e7c-7cac-492a-937e-32897fd3dbe6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I91837f49be6be9e5261496545eb4c26c2bb4e8cf
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com