Revert "Replace BTM_BleIsLinkKeyKnown with BTM_IsBonded"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ed0c4118-5d7c-476f-aa94-d86cf8a67306). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5229626f467e50088bb6fd8a1c647e449a84ac24
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com