Revert "Add flag scan_controller_thread"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bc34e163-2c02-48a7-8cc4-55e4a0990081). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If1fa3a07b4fc95b81cd629491aef408f6943057f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com