Revert "LeCallControl: delete dead code"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1888815c-7f7c-4c9e-b272-0ccc9628e8c1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 400845401
Change-Id: If850239406ab4b813cddeea577bb7afa0bee784d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com