Revert "LeCallControl: define internal value... internally"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1888815c-7f7c-4c9e-b272-0ccc9628e8c1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 400845401
Change-Id: I4be2a9f3626f5bce40655503ec9fa4cf3e62604a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com