Revert "floss: Fix the battery level update in the BAS"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1a927dc8-6c9a-4b3d-bb8e-8caff4721264). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia98a66f10b98e7ef01a85408a1d4291ad8193d64
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com