Revert "Errorprone fix & enforce FieldCanBeLocal"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4d13f5ea-5d57-4d37-b21f-7a75ad27ec54). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 400846416
Change-Id: I82e44145868069276137e4371a47d472d1d7738c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com