Revert "LeCallControl: define internal value... internally"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4d13f5ea-5d57-4d37-b21f-7a75ad27ec54). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 400846416
Change-Id: Ib1cfe7e8b96cf9ca1da8afd585d5f53c1431a793
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com