Revert "leaudio: Fix handling AvailableContext changes during st..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f6d8da7a-de6f-4be3-8173-f6f5687beb26). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9d4318c03ca8c6ff11d3d754dc1bff99e434b312
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com