Revert "Flag for allowing early incoming HID connection"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9d443a86-f5f5-4594-9c63-9e24552e83d6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I11d647b0989c6a8071160d49975b8d645237b5a0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com