Revert "Remove dependency from system/stack/test on system/gd"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d785c388-2165-497d-abb9-65afaca34ace). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0592a9ebdc99a0b87ab9a237e6ae8f7ca86e240d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com