Revert "Log actual requested tx_pdu_length and tx_time in BTM_Se..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a2ced2e5-141b-4184-acd3-4411ce24c0f6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4622e9b3efb18d6ebea4328851ea6e58ed024d4f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com