Revert "Run all distance measurement operations (Java) on the sa..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/83af6dee-3b39-4519-86e5-cbf4e137f64f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1a078c35e699e349bb5d232d26071712efde625f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com