Revert "Remove multiple media player support for cars."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2fa0dd1d-6548-4c87-9a63-83a9ef9b026d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I09f33a4e46468060b3d66185d332b97a8d8baa9a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com