Revert "Optimize and clean up DatabaseManager"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/89023f6f-8aa9-44f7-9453-f33c81d008bc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3bd780829312c4a3b4c87b412588e1c045294250
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com