Revert "Optimize and clean up DatabaseManager"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/29168227-b8cf-429f-ae0c-e575ce9eef0f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I31555f3e6d211f59904140b04596b1ef904606e2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com