Revert "Remove multiple media player support for cars."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dd243661-c802-4034-a629-4e4748ffc1fb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7eb2a8816c3abf8da4d33afc98c414615aeef624
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com