Revert "le_audio: state_machine: Ignore unexpected CIS establish..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f6e18509-d25a-4a38-9f4e-f2be07cac71e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6fa1648d0d0961dc90563f95b801b2ac19f560f6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com