Revert "le_audio: Wait for all CISes and Data Paths to be complete"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f6e18509-d25a-4a38-9f4e-f2be07cac71e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9f358742e3ba63544620d90019c1650625a32274
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com