Revert "Remove multiple media player support for cars."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/29862540-ac3c-4cd1-b586-0232017a0937). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia10200723f2c0e64eb2db355df490a491b121151
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com