Revert "Add tolerance for warning log of address rotation timing"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/41cc7eeb-0252-4604-b540-77ebccfaa4de). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1169cd71b7b25b8694975229470805418bd1fba9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com