Revert "Remove multiple media player support for cars."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/132c69e3-57b4-495c-8724-4b42f9bfc721). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I07efed86e41dc7e0dffe28ed93e264726121b552
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com