Revert "Remove multiple media player support for cars."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/802e56e6-4ce7-402f-b5b1-e9b12a0eec7f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6f3a239cd6269ca1151daa9d5a5b5b0fb9efc219
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com