Revert "Cleanup python version properties"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2719a519-6afd-4d0a-899b-c5316e9ca92f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I215088a3d837a2685561bf2e71ecbf9d0d25e69a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com