Revert "Add tolerance for warning log of address rotation timing"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f768101e-afc1-4236-9e80-4d4204adeadf). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iaa8266e89e90d83ae3a7f3aee1daa34d278cd6a5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com