Revert "Add flag scan_controller_thread"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/66e8c792-e560-4129-ae9d-f816fce1971b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id5f347fc906a91567df447d8f37d145785c8feb5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com