Revert "Add tolerance for warning log of address rotation timing"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3bf17727-d21d-46e1-ada6-afe7f1899a23). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I82c8476660acb871aa51a6c774522725d171cf23
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com