Revert "le_audio: state_machine: Ignore unexpected CIS establish..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4b7d5f4e-c692-416e-a1d7-ba5d67e9be6e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4a57bd2ead80909f84351408d5433136919c4c0d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com