Revert "Flag for resetting security flags on pairing failure"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5d3e9930-f71a-4830-8c97-23198766ad0b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib40cc7a71b0e348c06cfce8f203f5f381a44ebca
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com