Revert "le_audio: state_machine: Ignore unexpected CIS establish..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8f775fb7-2b1f-4000-bdb2-a217890c2cd3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0ed384887829fcdb73ebf9e3d14282b500a42fea
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com