Revert "le_audio: state_machine: Ignore unexpected CIS establish..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8f775fb7-2b1f-4000-bdb2-a217890c2cd3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iec81339e2ec70012e372f2a1431ebf17f8138420
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com