Revert "le_audio: Wait for all CISes and Data Paths to be complete"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8f775fb7-2b1f-4000-bdb2-a217890c2cd3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iaa56b5f98095581d9bd5b97dece2f707a69d09b9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com