Revert "le_audio: state_machine: Ignore unexpected CIS establish..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e49da8b1-805c-46fc-a8c5-9214ed1de70c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I27063b4cc5dbde519d86770c5e385a00f192cfaa
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com