Revert "le_audio: Wait for all CISes and Data Paths to be complete"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e49da8b1-805c-46fc-a8c5-9214ed1de70c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I45d62c7ca1369a656f5ec1a160b2a91794c61e47
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com