Revert "Add flag scan_controller_thread"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5257e8af-e4be-4202-96ca-7b6be527475d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6dee9e2b31f7b2d159a8b9ee1330d127b3ae9699
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com