Revert "Add tolerance for warning log of address rotation timing"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1b55202f-724a-4317-9d21-9243e69ceb66). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ica3cd3abc4809bd79befc3c91869e862909aa4ce
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com