Revert "le_audio: Wait for all CISes and Data Paths to be complete"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bb1d7098-a119-43fc-8aed-3522d1da5d6b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I22e17285f67500cdfb6d4629cc04a70927f5a411
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com