Revert "le_audio: state_machine: Ignore unexpected CIS establish..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bb1d7098-a119-43fc-8aed-3522d1da5d6b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3f373ff454718de6e601b2c7941afb5e854032b6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com