Revert "Reset BTM SEC security flags when device pairing fails"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e8983ed6-133c-46f2-b056-59522d2c59c1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0fa3510ce74b1eaa3646ad1a09b5f3ae90b59f18
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com