Revert "Log actual requested tx_pdu_length and tx_time in BTM_Se..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4c0c4ad3-c3f0-4756-a57e-ee918fd43e8f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I180fce94e85e1853480b6ee96edddbee13bd548e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com