Revert "le_audio: Wait for all CISes and Data Paths to be complete"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a4ad09e1-7aae-44c9-ae4f-f6ba19e75c58). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3012e9245d285c629090a2c459e49f12fd0d00b5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com