Revert "le_audio: state_machine: Ignore unexpected CIS establish..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a4ad09e1-7aae-44c9-ae4f-f6ba19e75c58). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I81c552e8475f88d58f16857f32862f130c3075f1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com