Revert "Remove multiple media player support for cars."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7638d614-cccd-48eb-b872-0124ee61e872). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic496299b60b0350bc83d187ba7b7098279279690
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com