Revert "Remove multiple media player support for cars."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4a0123d9-e40a-49d0-af38-03a5890c1b76). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I55d64f03afdc5a3c930d8cf927697836c2c68025
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com