Revert "Remove multiple media player support for cars."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f9cfe2e0-7610-410a-8845-a262f9050148). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I97cb7e05a130ffec7938acf7b6645c191c69c9b6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com