Fake Incoming Call Indicator after Active Call ends
Some remote devices are misbehaving when there is an active
call and MT call is in progress. We send call indicators for
active call and MT Call in progess. When active call is
terminated, we send call end indicator only to remote. This
is confusing remote and remotes are going into bad state.
Blacklist remote devices to disconnect SCO when active call is
ended, fake MT call indicator again with some delay.

Change-Id: Ic6ac7e6eb9743e19538b61e95d27f157c7586188