Revert "Remove multiple media player support for cars."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7064a79c-e9d1-4c0f-9ab5-36439dda5ccf). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I01b12d793cf9ec345d69eee46ea3a91d82ab57da
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com