Revert "Remove multiple media player support for cars."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9bb9076c-bcd5-4b24-a1a7-3aa92805715a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9df058b096807d23ccae2d67c788e834539186af
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com