Revert "Remove multiple media player support for cars."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2bead602-31be-4122-aada-1424c2cf042a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8cb2f910a31d1d12d7c3ca7fd628d8ef6cc3757c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com