Revert "Re-enable hfp tests now that we can do final mock"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f7a165cf-7af4-41e3-a747-18b618446690). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If43148fdb66a00723dcc0111b9f3f6ae99791bf3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com