Revert "Track and set the pairing persistency for all BR/EDR pairing"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f7a165cf-7af4-41e3-a747-18b618446690). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6ba8c153126321b8628ae522475cac47e22998de
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com