BCS: Add change to pass subevent timestamp in nano seconds to HAL.
steps:
1. Start BCS mode 3 ranging between initiator and reflector.
2. Start walk distance measurement, increasing the distance between the
   devices gradually.

[Expected Result]
No dips in distance measurement should be observed.

[Observed Result]
Dip in distance measurements at 60m distance.

Repro rate: Not mentioned (Every time ig)

Root Cause:
Host is not filling timestamp nanos value on Ranging AIDL before passing
it to HAL. This will cause dip in distances when calculating in
Algorithm.

Solution: Calculate timestamp difference between start of procedure and
when we pass values to HAL in nanoseconds. Pass this calculated value to
Algorithm to avoid dips and inconsistencies in distance.

Change-Id: I9433c8d59824b186eb4bd3233fad07d2279776f9