Correct call state in Conference call constructor
If there is conference call on phone in held state and then head-unit
and phone are connected over BT HFP, conference call becomes active.

Cause: From the constructor of HfpClientConference, conference call
state is set active. When we get onConnectionAdded() from telecom,
based on the conference call state (which is already set active in
constructor), we call connection.onAnswer() which will send AT+CHLD=2
to phone and call on phone side becomes active.
For second onConnectionAdded() also, another AT+CHLD=2 is sent to
phone which results two times AT+CHLD=2 to phone. So, sometimes, this
conference call becomes held->active and sometimes held->active->held.

Fix: A new parameter is added in HfpClientConference constructor which
is about the call state. If calls are getting merged from head-unit, it
will be default Connection.STATE_ACTIVE. If calls are merged from phone
or already active/held conference call is ongoing on phone, in that case
while contructing HfpClientConferecne, we will pass the parameter as
the state of the call being added in conference.

Bug: 217894838
Test: Manual
Change-Id: I345197ce535324e32d5f3aa6c42e703c313a3b81