osi: compat: fix memset builds with clang+glibc >=2.41
The BT tree is used by ChromeOS where it gets built
using clang and glibc. Starting with clang 21 and glibc
2.41 warnings like the following are emitted for memset
and are treated as errors:

/bta/hf_client/bta_hf_client_main.cc:253:10: error: first argument in call to 'memset' is a pointer to non-trivially copyable type 'tBTA_HF_CLIENT_CB_ARR' [-Werror,-Wnontrivial-memcall]
  253 |   memset(&bta_hf_client_cb_arr, 0, sizeof(tBTA_HF_CLIENT_CB_ARR));
      |          ^
bta/hf_client/bta_hf_client_main.cc:253:10: note: explicitly cast the pointer to silence this warning
  253 |   memset(&bta_hf_client_cb_arr, 0, sizeof(tBTA_HF_CLIENT_CB_ARR));
      |          ^
      |          (void*)
1 error generated.

Those in the cleanup functions for simple structures
are ok, the one in hf_client can be replaced with a
simple value init, however the one in btif is more
complex so I added a TODO to eventually write some
proper cleanup logic.

Bug: 324437695
Test: m com.android.bt; clang+glibc build in ChromeOS.
Flags: Execption, trivial
Change-Id: I159950f1a8ad9d1ddd6588abca3e74004510a6a0