BCS: Restart procedures if difference between local and remote procedure counter exceeds procedure data buffer size.
Steps:
1. Start music through A2DP profile on intitiator side.
2. Start distance measurment between initiator and reflector.

[Expected Result]
Channel Sounding should work seamlessly.

[Observed Result]
Distance Measurement stops after running for some time.

Repro rate: Not mentioned.

Root Cause:
 Issue is because of delay in receiving the remote data over the GATT to local DUT due to concurrent scenario.
So on the local and remote devices, the procedure data is getting stuck in the queue, and we are processing them little later.
After reaching the threshold both the DUT and Remote are getting out of sync and we are not able to recover from that state.

Solution:
Apply procedure restart logic, in scenarios when there is delay in receiving RAS packets and local and remote data go out of sync.
 After procedure is restarted, the data should be in sync and distance measurement should work normally.

bug:441192189

Change-Id: I4c1b680f9207272ea08dc1dfb508f5e9e98b2044