Reset server congestion to false if GATT is disconnected
The DUT initiates data transmission over the ATT channel, during which the congestion flag is set to true. Subsequently, the DUT disconnects the GATT link, but the congestion flag remains true and is not reset. Upon reconnection, the DUT attempts to resume data transmission over the ATT channel but fails, as the lingering congestion state prevents successful communication.

Reset server congestion to false if GATT is disconnected

Test: Gatt connect/disconnect/reconnect Flag: EXEMPT, trivial

Bug: 449615796
Change-Id: I5ec2eea7c042e18880a9abd708b96292d41b2e50