[BLE] BUG_FUNCTION_20
During an L2c link state is disconnecting state(Remote disconnection case) and a COC connection is requested, an automatic reconnection is initiated by disconnection_complete event.
During this, the lcb is reused. If this automatic reconnection fails to establish and times out, the lcb is not released. 
This issue occurs only if the holding application deregisters or close before the l2c timer timeout, which cancels the l2c timer. (Normally, the lcb would be released after the l2c timeout)
Subsequently, if the BR attmpts to reuse the same ACL problems such as SDP failures may arise.

Bug: https://partnerissuetracker.corp.google.com/issues/429922295
Change-Id: Ie2723bff0736b8c8d1471359b8091c0962264243