LE: Fetching the Remote version info from BREDR control block
Issue:
During LE connection, when we are trying to check Robust caching
supported or not. It needs to check for remote version but
on LE transport remote version command is send and waiting
for remove version event. So it will return 0 if we checks before
event received even though its already fetched in BREDR transport.

Fix:
If the BREDR transport is already connected and remote version
is already fetched, then we dont need to send remote version
command and wait for the response event. We can reuse the version
information from BREDR tranport.

Change-Id: Ieffb2677f25fa8a781a76cc2edc6703da6ba25a9