Existing conference call on AG not handled on HU
Cause: If phone has existing conference call and connects with HU,
telecom api is called to add call with telecom and at the same time,
hfpclient tries to add the call in conference. Hfpclient skip adding
call into conference as call is not added with telecom by that time.

Fix: When onCreateUnknownConnection is received from telecom, then
hfpclient will attempt to add the call into conference.

Bug: 366190459
Test: m com.android.btservices
Test: Manual

Change-Id: Idff17c902ae4523e943570330e0ad9d11b33cef0
Signed-off-by: Kapil Gupta <Kapil.1.Gupta@gm.com>