HID: Set hh_poll_thread_id to invalid handle before joining poll thread
bta_hh_co_close can be invoked from multiple threads, and in some cases an invalid thread
ID is passed, which leads to an assert in pthread_join. The poll interval is 50 ms, so
the poll thread will be destroyed within 50 ms after the poll is removed. To avoid joining
a stale or invalid thread, set p_dev->hh_poll_thread_id to an invalid handle before calling
pthread_join.

Bug: 463533290

Test: Bluetooth Compilation and HID stability use case.
Change-Id: I50f08055176c097cf2e5cee6205413f4a3704291