Fix remove bond fail
[Root Cause]
Fix the remove bond failure: When the remove bond operation and the reconnection operation occur almost simultaneously, the remove bond operation sometimes fails. If during an RC reconnection the ACL is connected but the HCI_Read_Remote_Version_Information_Complete event is not received from the firmware, the flag p_acl.link_up_issued is not set to true. As a result, when remove bond attempts to disconnect, the function NotifyAclLinkDown will not call BTA_dm_acl_down, causing the remove bond operation to not complete

[Solution]
If The flag pacl.link_up_issued is no set true, NotifyAcnLinkDown
need to call bta_dm_remove_on_disconnect

Bug:
Test: atest net_test_bta pass
Change-Id: I11ceb9519b3eabfe17daffc2d443ae1835a184fb